requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-database-url==2.1.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
